+++ essencium-backend/src/main/java/de/frachtwerk/essencium/backend/model/Role.java
@@ -75,6 +75,6 @@ public class Role implements GrantedAuthority {
 
   @Override
   public int hashCode() {
-    return Objects.hash(getName(), getDescription(), isProtected(), getRights());
+    return Objects.hash(getName());
   }
 }